+++ Asset Management System/Database/Repositories/IRepository.cs
@@ -7,7 +7,7 @@ namespace Asset_Management_System.Database.Repositories
     public interface IRepository<T>
     {
         public void Insert(T entity);
-        public T GetById(int id);
+        public T GetById(long id);
         public void Delete(T entity);
     }
 }
